+++ frontend/react/src/App.css
@@ -590,7 +590,3 @@ img {
 .percentages .divide {
   margin: 16px;
 }
-
-.ds-c-table-borderless {
-  border: none;
-}
