+++ frontend/react/src/components/layout/Header.js
@@ -53,7 +53,7 @@ class Header extends Component {
                   </ul>
                   <ul className="menu-block" id="menu-block">
                     <li className="manage-account">
-                      <a href="../../profileInformation">Manage account</a>
+                      <a href="/user/profile">Manage account</a>
                     </li>
                     <li className="logout">
                       <Logout />
