+++ frontend/react/src/components/sections/UserProfile.js
@@ -1,0 +1,56 @@
+import React, { Component } from "react";
+import { connect } from "react-redux";
+
+class UserProfile extends Component {
+  constructor() {
+    super();
+  }
+
+  render() {
+    return (
+      <div className="page-info">
+        <div className="ds-l-col--12 content ds-u-padding-left--4 ">
+          <h1>User Profile</h1>
+          <div className="main">
+            If any information is incorrect, please contact the{" "}
+            <a href="mailto:cartshelp@cms.hhs.gov">CARTS Help Desk</a>.
+            <div className="profile-information ds-l-container">
+              <div clasName="ds-l-row">
+                <div className="ds-l-col--4">EUA Id: </div>
+                <div className="ds-l-col--8">
+                  {this.props.currentUser.username}
+                </div>
+              </div>
+              <div clasName="ds-l-row">
+                <div className="ds-l-col--4">Name: </div>
+                <div className="ds-l-col--8">Full name here</div>
+              </div>
+              <div>
+                <div className="ds-l-col--4">Email: </div>
+                <div className="ds-l-col--8">
+                  {this.props.currentUser.email}
+                </div>
+              </div>
+              <div>
+                <div className="ds-l-col--4">State: </div>
+                <div className="ds-l-col--8">
+                  {this.props.currentUser.state.name}
+                </div>
+              </div>
+              <div>
+                <div className="ds-l-col--4">Role: </div>
+                <div className="ds-l-col--8">{this.props.currentUser.role}</div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
+}
+
+const mapStateToProps = (state) => ({
+  currentUser: state.stateUser.currentUser,
+});
+
+export default connect(mapStateToProps)(UserProfile);
