+++ frontend/react/src/reactRouter.js
@@ -8,7 +8,7 @@ import {
 } from "react-router-dom";
 import { constructIdFromYearSectionAndSubsection } from "./store/formData";
 import Homepage from "./components/sections/homepage/Homepage";
-import ProfileInformation from "./components/sections/ProfileInformation";
+import UserProfile from "./components/sections/UserProfile";
 import Review from "./components/review/Review";
 import Sidebar from "./components/layout/Sidebar";
 import Section from "./components/layout/Section";
@@ -33,7 +33,7 @@ const Routes = ({ userData }) => (
         <ScrollToTop />
         <Switch>
           <Route exact path="/" component={Homepage} />
-          <Route path="/profileInformation" component={ProfileInformation} />
+          <Route path="/user/profile" component={UserProfile} />
           <Route path="/reports/:stateAbbrev/:year" component={Review} />
           <Route path="/sections/:year/:sectionOrdinal/:subsectionMarker">
             <InvokeSection userData={userData} />
